Add the groovy
This is a direct lift from our Puppet repository.

Splitting this out to it's own repository makes sense in that it is
easier to maintain and understand within version control and history; it
can be loaded into Jenkins instances without requiring our full Puppet
repository; and we could potentially add some basic unit tests to the
Groovy code.