Add Jenkinsfile
This Jenkinsfile could be used to run library unit tests in the future,
but for now it just ensures it can load the library.

It's slightly tougher to test the changes inside it. You can create a
new branch, and specify it in the Jenkinsfile itself:

`library("govuk@my-branch-name")`

That would then hopefully load the branch that you wish to test.