Run brakeman during builds
The brakeman gem will analyse the application and warn of any security
issues. To avoid breaking the tests for all applications the script
reports the result of the security check separately to GitHub. This
will allow us to fix any issues without disrupting the workflow. But in
a couple of weeks we'll require the security check to pass in order to
merge the PR.