Add Shellcheck method
This uses the shellcheck tool to test shell scripts.

The method will default check *.sh files, and in this mode exclusions
may be passed in to ignore false positives.

It is also possible to pass in a specific list of files, but passing in
a specific list and exclusions cannot be used together.

This kind of check should be specifically added for a repository.