Support not passing --diff to govuk-lint-ruby
It's annoying when making a change to have the linter complain about
something unrelated, but on the same lines. This is due to the --diff
option. When it's used, changes to the linter or linting rules may
make some of the existing code invalid, but because those lines
haven't been changed, it isn't picked up by CI.

Therefore, allow disabling the `--diff` flag by passing in a
`rubyLintDiff` option.