Set a fake SECRET_KEY_BASE when precompiling assets
This commit sets a fake `SECRET_KEY_BASE` when precompiling assets. This is because starting with Rails 5.2, `SECRET_KEY_BASE` is required when running any command, and CI doesn't have access to this value when running with `RAILS_ENV=production`. This worked previously since a missing `SECRET_KEY_BASE` would only generate a warning rather than an exception.