Support disabling Docker related tasks
Currently these run by defualt if a Dockerfile exists, but this
results in doing unnecessary work for some repositories that contain a
Dockerfile, but one which isn't used (the seal is one example). This
should make it possible to disable the Docker related tasks, to avoid
the associated extra work and potential failures.