Create Mongoid indexes when running tests
For an application that uses Mongoid, it can sometimes be necessary for the indexes to be there. Specifically, this is the case when using the geospatial modules https://docs.mongodb.com/manual/geospatial-queries/ where there has to be an index to perform these queries.

This is currently making the imminence tests flaky, as it depends on which `ci-agent` machine they run on, since only some of the databases have had those geospatial indexes built.