Lint all the code by default
Rather than just checking what has changed. This works much better for
updates of the linting rules, as it means that Jenkins will check all
the code, ensuring that any issues are raised.

I've been going around and fixing linting issues in various
repositories, so I think as it's firebreak week, we can have a go at
changing the default (and fixing issues as and when they arise).