Workaround merge issues with shallow clones
There's some value in merging the master branch, as this is a slightly
more rigorous test of the state that would result from merging the
branch. There's also value in only doing a shallow clone, as this
mitigates the space and time issues of constantly growing Git
repositories.

Currently, merging master is failing as Git doesn't have enough
information to perform the merge. I think if a newer version of Git
was in use, it would even refuse to attempt the merge.

Therefore, to try and get the best of both worlds, fetch more commits,
so that merging master works for most Pull Requests.