Precompile assets before testing
Move the precompile asset step before the testing step. Also set the
`RAILS_ENV=test` for asset precompilation.

In some of our applications such as government-frontend, we rely on
Capybara and Selenium to do some front end testing. When Capybara kicks
off, it tries to precompile assets if they aren't already precompiled.
This can take quite a while and sometimes causes `Net::ReadTimeout`
errors. If we instead precompile assets before Capybara runs, it will
use the precompiled assets and skip trying to precompile them. This
means we don't spend time during the first request waiting for
precompilation and we are less likely to hit the timeout.

Changing `RAILS_ENV` from `production` to `test` so that during testing
we use the precompiled assets. This shouldn't affect behaviour because
the intent of the asset precompilation step is test for syntax errors
in the CSS which will still happen.