Remove diff linting functionality
This removes the option that only the diff of Ruby code can be linted.

We started the process to stop only linting diffs back in May 2018 with
https://github.com/alphagov/govuk-jenkinslib/pull/9 when the option to
turn off diffing was first created. An attempt was made to make this the
default was done in https://github.com/alphagov/govuk-jenkinslib/pull/35
however this was reverted in
https://github.com/alphagov/govuk-jenkinslib/pull/36 since it broke
quite a few repos.

Nearly a year later we're more confident as the roll out of govuk-lint
4.0 https://github.com/alphagov/govuk-lint/pull/126 has meant a
significant amount of repos are now no longer diff linting and have
meet current lint requirements.

There will no doubt still be a few things that fail due to this - I'm
sorry - but this will help us move towards the world where we can
consider adopting the outcome of
https://github.com/alphagov/govuk-rfcs/blob/master/rfc-100-linting.md

I decided rather than to make this an option we could just turn it off.
For apps that fail because of this there is an expectation that linting
is fixed rather than allowing them to use a deprecated behaviour. If
this causes people misery the operation can return but defaulted to
false.