Bundle exec brakeman if it's available
This simplifies how we run Brakeman to use the version specified and
installed by Bundler. Mostly this means builds are more reproducible
as we always use the version specified in the `Gemfile.lock` rather
than the version installed on CI.

I'm also hoping this will solve a problem we have with intermittent
Brakeman runs where it doesn't seem to install properly.