Set-up for handling non-master default branches
This allows determining the default branch from the Git repo directly.
It then tries to update the many references to master to be parameters.

This was a bit of a pain to do because all the functions in here may be
used by other GOV.UK projects so should be backwards compatible. This
has made it difficult to do any refactoring and made all these changes
rather convoluted.

A non-backwards compatible change is that deployIntegration no longer
has a guard clause in it. As far as I can tell this is always executed
within a conditional anyway so this doesn't seem of consequence - it
also seems a bit bizarre to have a function that is given a branch as a
parameter and then guard clauses on that only having one value.