Implement default branch detection
Use `git ls-remote` to get the commit hash of the HEAD ref, and check
whether a given branch name matches it.  We then use this to choose
between "main" and "master".

`git ls-remote --symref` would be a better way of doing this, but the
version of git on the ci-agents is too old and does not have that
flag.