Remove branch argument from deployIntegration
To maintain backwards compatibility I've renamed the function to
`deployToIntegration` and left the old `deployIntegration` function
which calls the new one but drops the `branch` argument.

The clause has been removed from the `deployToIntegration` function
because as far as I can tell this is always executed within a
conditional anyway so this doesn't seem of consequence - it also seems
a bit bizarre to have a function that is given a branch as a parameter
and then guard clauses on that only having one value.