Set-up for handling non-master default branches
This allows determining the default branch from the Git repo directly.
It then tries to update the many references to master to be parameters.

This was a bit of a pain to do because all the functions in here may be
used by other GOV.UK projects so should be backwards compatible. This
has made it difficult to do any refactoring and made all these changes
rather convoluted.