Fix broken interpolation
The single quotes don't allow for interpolation which meant the $branch was
treated as undefined shell variable. Switching to double quotes fixes
this and resolves this error we were seeing:

```
+ git rev-list origin/
fatal: ambiguous argument 'origin/': unknown revision or path not in the working tree.
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'
+ git rev-parse HEAD
+ grep 7dd2357301eb757a2e4098295dfcfa528c360851
[Pipeline] echo
Current commit is not on main, so attempting merge of main branch before proceeding with build
```