Use single quotes to wrap s3cmd in shell step
This commit replaces double quotes for single quotes to wrap the `s3cmd` which is passed to the `sh`ell step, as [recommended by the Jenkins docs](https://www.jenkins.io/doc/book/pipeline/jenkinsfile/#interpolation-of-sensitive-environment-variables) to avoid printing out sensitive data to logs. This manifests in our CI builds as a warning on the job status.