Accept `ARG ruby_version=` as an alternative to `FROM ruby:`.
We want to use a better base image than the community-supported Ruby
ones, so that we can receive security updates promptly. This test was
erronously asserting that we use the community-supported image.

Alternatives considered:

I started eliminating this nasty [change-detector] test by simply
passing the Ruby version from `.ruby-version` to `docker build`, but
then realised it'd be a lot of work to test given that we need to keep
supporting the old Dockerfiles for the time being. The test will become
obsolete anyway once we've given all the Dockerfiles the Replatforming
treatment.

[change-detector]: https://testing.googleblog.com/2015/01/testing-on-toilet-change-detector-tests.html