Use new Docker Enterprise user
We have a new user `govukdockerci` which has been added to our
Enterprise hub and is an owner of the `govuk` organisation.
This should allow us to push without any limits. We will continue
to push to the existing govuk organisation that is owned by the
`govukci` user.

This avoids the disruption of having to recreate all the repos in
the governmentdigitalservice org which is shared across GDS.

As I understand we only use Dockerhub as a way to share images with
publishing-e2e-tests. So once these are removed [1] this quirk can be
removed.

[1]: https://github.com/alphagov/publishing-e2e-tests/commit/f0d0d56185cf0bb4bd90c83ec78d085e25032de7

Co-authored-by: Kevin Dew <kevin.dew@digital.cabinet-office.gov.uk>