Replace deprecated commands with bundler suggestions
This change resolves two things:

1) It fixes a problem with newer versions of bundler that have a feature
   of installing the requested version of bundler [1] and re-running
   with that version (using the old command causes a write access issue:
   Bundler::SudoNotPermittedError: Bundler requires sudo access to install at
   the moment. Try installing again, granting Bundler sudo access when
   prompted, or installing into a different path.)
2) It stops each build outputting deprecation warnings such as
   `[DEPRECATED] The `--deployment` flag is deprecated because it relies
   on being remembered across bundler invocations, which bundler will no
   longer do in future versions. Instead please use `bundle config set
   --local deployment 'true'`, and stop using this flag`

The known downside to doing this is that we lose the ability to run old
versions of bundle that don't support this argument. However this
concern should only be a problem for software running on EOL versions of
Ruby - as the earliest non-EOL version of Ruby
(2.7) is packaged with Bundler 2.1.2 and these deprecations were
introduced in Bundler 2.1.0.