Configure Bundler based on version used
Bundler has deprecated passing the without, path and deployment
arguments to bundle install. These will be removed in Bundler 3.

We also found that, when using a newer Bundler feature: auto install
[1], in our Ruby 3.1 builds that `bundle install` can fail with
permissions errors. I applied a change to fix this in adc5e6a.

However in that fix I overlooked that we actually use an extremely old
version of Ruby for govuk-puppet, 1.9.3 😭 - which is unlikely to change
soon.

So to resolve this I've gone for the rather complex change of adding a
conditional to check which version of Bundler is used. This meant the
dreaded process of doing a semvar comparison. I'm not knowledgable
enough to know if there is anything in Groovy's std library that could
make this easier, so I took a crude approach.

There were a couple of alternative approaches that I rejected.

1) Set long lived Bundler environment variables. As far as I can tell the same
   environment variables will work with Bundler versions from 1.x to
   current 2.x. I chose against that as these would apply beyond the
   directory of the bundling and so could cause problems if any
   Jenkinsfile calls both the bundleGem and bundleApp.
2) Set short lived environment variables. We could have done
   `BUNDLER_PATH=/path bundle install` to not have env vars affect other
   calls. However this would not be compatible with Bundler 3 which
   would need these set for commands like `bundle exec`.

I've done a couple of example builds to demonstrate this running
successfully:

- publishing-api on Ruby 3.1.2: https://ci.integration.publishing.service.gov.uk/job/publishing-api/job/ruby-3.1.2-again/20/
- govuk-puppet: https://ci.integration.publishing.service.gov.uk/job/govuk-puppet/job/test-jenkinslib/4

[1]: https://bundler.io/blog/2022/01/23/bundler-v2-3.html