Only use the governmentdigitalservice docker hub account and drop the govuk one
We're consolidating all of GDS' Docker account usage so that it goes through governmentdigitalservice.
This org has multiple owners and a process for requesting additional seats, etc, unlike the govuk org which has a 3 seat limit.

We started building two images and pushing to both accounts in https://github.com/alphagov/govuk-jenkinslib/pull/113

This removes the build and push steps to the govuk account.

https://trello.com/c/6JRYK8hU/2982-push-to-pull-from-governmentdigitalservice-docker-org-3