Make shallow clones a config option that defaults to false
This has been done to prevent the common headaches devs experience when
they have a large PR, or collection of PRs that end up being more than
30 commits out of sync with the main branch.

In those cases developers are greeted with a rather unhelpful message
of:

```
+ git merge --no-commit origin/main
fatal: refusing to merge unrelated histories
+ git merge --abort
fatal: There is no merge to abort (MERGE_HEAD missing).
```

As we don't have an agreement that there is a commit limit on PRs this
arbitrary number is confusing and divisive (easily starts the debate on
how big should a PR be).

This changes the approach to default to not having any limit at all and
makes the limit an optimisation that repos can opt into. This means that
repos calling `govuk.buildProject` will default to cloning an entire
repo.

My theory is that there is actually very few GOV.UK repos where the
quantity of commits is significant enough that the time spent cloning
repos is a noticeable impact in build times (the worst one was
govuk-secrets which is now on GitHub Actions). If there is a repo that
is big the team owners can configure their call to buildProject to
configure their project accordingly.

For example to restore the previous behaviour to a repo you can call
govuk.buildProject like so:

```
govuk.buildProject(shallowClone: env.BRANCH_NAME != "main", mergeDepth: 30)
```