Remove Publishing E2E tests from CI pipelines
We're decommissioning the [Publishing E2E tests][] now that GOV.UK apps have transitioned to a model of continuous deployment with contract tests, as outlined in [RFC-128][]. The overall progress of this work is tracked in a [Trello card on the GOV.UK Tech Debt board][tech-debt-card].

This commit removes functionality from the `govuk.buildProject()` method that triggers the Publishing E2E tests to run against PRs in app repos. It stops the option `publishingE2ETests: true` from having any effect in the CI pipeline, as we no longer need them to run.

A [separate card exists][jenkinsfile-card] to tidy up the Jenkinsfile for each affected application.

[Publishing E2E tests]: https://github.com/alphagov/publishing-e2e-tests
[RFC-128]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-128-continuous-deployment.md#delete-publishing-e2e-tests
[tech-debt-card]: https://trello.com/c/Lbw4TTfD/233-publishing-e2e-tests-still-exist
[jenkinsfile-card]: https://trello.com/c/qgR41OnR/837-disable-publishing-end-to-end-tests-on-ci-pipelines