Remove building Dockerfile from the buildProject() process
This change removes the building of a Dockerfile from the default
buildProject process. This has been done due to the retirement of the
publishing-e2e-tests which made use of these images [1].

There is a CI check that is lost due to this, the check that the Ruby
version in .ruby-version and the Dockerfile are the same version. I
talked to Chris Banks about this and the expectation is that the need
for this check will be removed by setting the ruby-version in the
Dockerfile with an argument. For example:

```
Dockerfile

ARG ruby_version
ARG base_image=ghcr.io/alphagov/govuk-ruby-base:${ruby_version}
ARG builder_image=ghcr.io/alphagov/govuk-ruby-builder:${ruby_version}
FROM $builder_image AS builder
```

and built with:

```
docker build . --build-arg ruby_version=$(cat .ruby-version)
```

This also removes the PUSH_TO_GCR and RUN_DOCKER_TASKS arguments as this
change removes their effects. No repositories were using them.

This doesn't remove all the Docker functions as there are still some
projects using them. They will be removed in a separate PR.

[1]: https://github.com/alphagov/govuk-jenkinslib/commit/e99c76c19fb2c815bcb75e0ea55eac2933d62bc8