Add ability to clone publishing api
This is part of work to move content schemas into publishing api, and should mean that apps can clone a specific branch of publishing api to test against. This will eventually replace the current behaviour which clones the govuk-content-schemas repo. For now, it should only happend when env var USE_PUBLISHING_API_FOR_SCHEMAS is set to true.